+++ packages/moonstone/Marquee/MarqueeDecorator.js
@@ -87,6 +87,17 @@ const didPropChange = (propList, prev, next) => {
 	return hasPropsChanged.indexOf(true) !== -1;
 };
 
+/*
+ * There's only one timer shared for Marquee so we need to keep track of what we may be using it
+ * for. We may need to clean up certain things as we move among states.
+ */
+const TimerState = {
+	CLEAR: 0,				// No timers pending
+	START_PENDING: 1,		// A start request is pending
+	RESET_PENDING: 2,		// Marquee finished, waiting for reset delay
+	SYNCSTART_PENDING: 3	// Waiting to alert Controller that we want to start marqueeing
+};
+
 /**
  * {@link moonstone/Marquee.MarqueeDecorator} is a Higher-order Component which makes
  * the Wrapped component's children marquee.
@@ -232,11 +243,13 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		constructor (props) {
 			super(props);
 			this.state = {
+				animating: false,
 				overflow: 'ellipsis',
 				rtl: null
 			};
 			this.sync = false;
 			this.forceRestartMarquee = false;
+			this.timerState = TimerState.CLEAR;
 
 			this.invalidateMetrics();
 
@@ -283,7 +296,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				this.calculateMetrics();
 			}
 			if (this.shouldStartMarquee()) {
-				this.startAnimation(this.props.marqueeOn === 'render' ? this.props.marqueeOnRenderDelay : this.props.marqueeDelay);
+				this.tryStartingAnimation(this.props.marqueeOn === 'render' ? this.props.marqueeOnRenderDelay : this.props.marqueeDelay);
 			}
 			this.forceRestartMarquee = false;
 		}
@@ -296,7 +309,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			}
 		}
 
-		/**
+		/*
 		 * Clears the timer
 		 *
 		 * @returns {undefined}
@@ -306,23 +319,28 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				window.clearTimeout(this.timer);
 				this.timer = null;
 			}
+			this.timerState = TimerState.CLEAR;
 		}
 
-		/**
+		/*
 		 * Starts a new timer
 		 *
 		 * @param {Function} fn   Callback
 		 * @param {Number}   time Delay in milliseconds
 		 * @returns {undefined}
 		 */
-		setTimeout (fn, time = 0) {
+		setTimeout (fn, time = 0, state = TimerState.CLEAR) {
 			this.clearTimeout();
 			if (window) {
-				this.timer = window.setTimeout(fn, time);
+				this.timerState = state;
+				this.timer = window.setTimeout(() => {
+					this.timerState = TimerState.CLEAR;
+					fn();
+				}, time);
 			}
 		}
 
-		/**
+		/*
 		 * Checks to see if the children changed during a condition that should cause us to re-check
 		 * the animation state
 		 *
@@ -340,7 +358,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			);
 		}
 
-		/**
+		/*
 		 * Invalidates marquee metrics requiring them to be recalculated
 		 *
 		 * @returns {undefined}
@@ -352,7 +370,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			this.contentFits = false;
 		}
 
-		/**
+		/*
 		* Determines if the component should marquee and the distance to animate
 		*
 		* @returns {undefined}
@@ -368,7 +386,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			}
 		}
 
-		/**
+		/*
 		 * Calculates the distance the marquee must travel to reveal all of the content
 		 *
 		 * @param	{DOMNode}	node	DOM Node to measure
@@ -381,7 +399,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			return distance;
 		}
 
-		/**
+		/*
 		 * Calculates the text overflow to use to correctly render the ellipsis. If the distance is
 		 * exactly 0, then the ellipsis is most likely hiding the content, and marquee does not need
 		 * to animate.
@@ -393,7 +411,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			return distance === 0 ? 'clip' : 'ellipsis';
 		}
 
-		/**
+		/*
 		 * Calculates if the marquee should animate
 		 *
 		 * @param	{Number}	distance	Amount of overflow in pixels
@@ -403,10 +421,10 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			return distance > 0;
 		}
 
-		/**
+		/*
 		 * Starts the animation without synchronizing
 		 *
-		 * @param	{Number}	[delay]	Milleseconds to wait before animating
+		 * @param	{Number}	[delay]	Milliseconds to wait before animating
 		 * @returns	{undefined}
 		 */
 		start = (delay = this.props.marqueeDelay) => {
@@ -416,6 +434,10 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				// instances.
 				return true;
 			} else if (!this.state.animating) {
+				// Don't need to worry about this.timerState because if we're sync, we were just
+				// told to start, so our state is correct already. If we're not sync, this will
+				// restart us anyhow. If we were waiting to tell sync to start us, someone else in
+				// our group already did it.
 				this.setTimeout(() => {
 					this.calculateMetrics();
 					if (!this.contentFits) {
@@ -425,11 +447,11 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 					} else if (this.sync) {
 						this.context.complete(this);
 					}
-				}, delay);
+				}, delay, TimerState.START_PENDING);
 			}
 		}
 
-		/**
+		/*
 		 * Stops the animation
 		 *
 		 * @returns	{undefined}
@@ -441,25 +463,44 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			});
 		}
 
-		/**
+		/*
+		 * Starts marquee animation with synchronization, if not already animating
+		 *
+		 * @param {Number} [delay] Milliseconds to wait before animating
+		 * @returns {undefined}
+		 */
+		tryStartingAnimation = (delay) => {
+			if (this.state.animating) return;
+
+			this.startAnimation(delay);
+		}
+
+		/*
 		 * Starts marquee animation with synchronization
 		 *
-		 * @param {Number} [delay] Milleseconds to wait before animating
+		 * @param {Number} [delay] Milliseconds to wait before animating
 		 * @returns {undefined}
 		 */
 		startAnimation = (delay) => {
-			if (this.state.animating) return;
-
 			if (this.sync) {
+				// If we're running a timer for anything, we should let that finish, unless it's
+				// another syncstart request.  We should probably check to see if the start request
+				// is further in the future than we are so we can choose the nearer one. But, we're
+				// assuming the condition is we're waiting on render delay and someone just hovered
+				// us, so we can start with the (hopefully) faster hover delay.
+				if (this.timerState !== TimerState.CLEAR &&
+						this.timerState !== TimerState.SYNCSTART_PENDING) {
+					return;
+				}
 				this.setTimeout(() => {
 					this.context.start();
-				}, delay);
+				}, delay, TimerState.SYNCSTART_PENDING);
 			} else {
 				this.start(delay);
 			}
 		}
 
-		/**
+		/*
 		 * Resets the marquee and restarts it after `marqueeDelay` millisecons.
 		 *
 		 * @returns {undefined}
@@ -472,21 +513,30 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			if (this.sync) {
 				this.context.complete(this);
 			} else {
-				this.startAnimation();
+				this.setState((prevState) => {
+					if (!prevState.animating) {
+						this.startAnimation();
+					}
+					return null;
+				});
 			}
 		}
 
-		/**
+		/*
 		 * Resets and restarts the marquee after `marqueeResetDelay` milliseconds
 		 *
 		 * @returns {undefined}
 		 */
 		resetAnimation = () => {
 			const marqueeResetDelay = Math.max(40, this.props.marqueeResetDelay);
-			this.setTimeout(this.restartAnimation, marqueeResetDelay);
+			// If we're already timing a start action, don't reset.  Start actions will clear us if
+			// sync.
+			if (this.timerState === TimerState.CLEAR) {
+				this.setTimeout(this.restartAnimation, marqueeResetDelay, TimerState.RESET_PENDING);
+			}
 		}
 
-		/**
+		/*
 		 * Cancels the marquee
 		 *
 		 * @returns {undefined}
@@ -507,7 +557,12 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 		handleFocus = (ev) => {
 			this.isFocused = true;
-			this.startAnimation();
+			this.setState((prevState) => {
+				if (!prevState.animating) {
+					this.startAnimation();
+				}
+				return null;
+			});
 			forwardFocus(ev, this.props);
 		}
 
@@ -519,7 +574,12 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 		handleEnter = (ev) => {
 			this.isHovered = true;
-			this.startAnimation();
+			this.setState((prevState) => {
+				if (!prevState.animating) {
+					this.startAnimation();
+				}
+				return null;
+			});
 			forwardEnter(ev, this.props);
 		}